requirements.txt
pyzmq
ultralytics[export]
opencv-python
Pillow
